davinci: dm646x: Add support for using calculated arm clock frequency for reference
The arm clock reference used is calculated from input clock value and PLL
register values. This helps using the same U-Boot binary on boards having
different clock input and/or where PLL values are manipulated to get overclocking.

The input clock is assumed 27MHz by default but can be overridden by setting
environment variable 'clkref' to desired frequency value. For example, on
DM6467T EVMs use 'setenv clkref 33000000' and then save the environment using
'saveenv'. The new clock reference value will be used ofter 'reset'.

An environment variable command 'addclk' is also provided which could be used to
append 'clkref' value to 'bootargs' to be passed to kernel. For example, if
'clkref' is set as '33000000', 'clkref=33000000' will be appended to 'bootargs'.
Note that this command should be used only when 'clkref' is set, otherwise a
blank value will be passed.