omap3: Fix MPU running at 381MHz
When s_init() is called, the silicon version hasn't been
identified leading to incorrect index in the DPLL table.

This patch ensures that silicon is identified as first
step in s_init().

When called from s_init(), the global variables seem to
be updated in a 'relocated' location. As a workaround,
this patch ensures that cpu_identify() is called again
in arch_cpu_init().